requirements.txt
# Core dependencies for the JP Morgan project
python-dotenv>=1.0.0
anthropic>=0.3.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0

# Optional dependencies for enhanced functionality
# yfinance>=0.1.70  # For financial data
# alpha-vantage>=2.3.1  # For additional financial data
# plotly>=5.0.0  # For interactive plots
